+++ apps/openmw/mwclass/potion.cpp
@@ -108,7 +108,7 @@ namespace MWClass
 
         // hide effects the player doesn't know about
         MWWorld::Ptr player = MWBase::Environment::get().getWorld()->getPlayerPtr();
-        for (unsigned int i = 0; i < info.effects.size(); ++i)
+        for (size_t i = 0; i < info.effects.size(); ++i)
             info.effects[i].mKnown = MWMechanics::Alchemy::knownEffect(i, player);
 
         info.isPotion = true;
